# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/plugins": {
      "post": {
        "operationId": "PluginsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePluginDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The plugin has been successfully created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PluginResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new plugin",
        "tags": ["plugins"]
      },
      "get": {
        "operationId": "PluginsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["active", "inactive", "deprecated", "removed"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all plugins.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PluginResponseDto" }
                }
              }
            }
          }
        },
        "summary": "Get all plugins",
        "tags": ["plugins"]
      }
    },
...
  },
  "info": {
    "title": "eTickets Platform API",
    "description": "The eTickets Platform API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreatePluginDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Plugin name",
            "example": "Stripe Payment Gateway"
          },
          "version": {
            "type": "string",
            "description": "Plugin version",
            "example": "1.0.0"
          },
          "description": {
            "type": "string",
            "description": "Plugin description",
            "example": "Integrates Stripe payment processing"
          },
          "category": {
            "type": "string",
            "description": "Plugin category",
            "enum": [
              "payment",
              "marketing",
              "analytics",
              "social",
              "ticketing",
              "layout",
              "seating"
            ]
          },
          "bundleUrl": {
            "type": "string",
            "description": "URL to the plugin bundle"
          },
          "extensionPoints": {
            "description": "List of extension points this plugin implements",
            "example": ["payment.gateway", "checkout.form"],
            "type": "array",
            "items": { "type": "string" }
          },
          "adminComponents": {
            "type": "object",
            "description": "Admin components provided by the plugin",
            "example": {
              "settings": "StripeSettings",
              "eventCreation": "StripeEventOptions",
              "dashboard": "StripeDashboard"
            }
          },
          "storefrontComponents": {
            "type": "object",
            "description": "Storefront components provided by the plugin",
            "example": {
              "checkout": "StripeCheckout",
              "widgets": { "paymentMethods": "StripePaymentMethods" }
            }
          },
          "metadata": {
            "type": "object",
            "description": "Metadata about the plugin",
            "example": {
              "priority": 100,
              "displayName": "Stripe Payments",
              "author": "Stripe Inc.",
              "paymentProvider": "stripe",
              "supportedMethods": ["credit_card", "ach"],
              "supportedCurrencies": ["USD", "EUR"]
            }
          },
          "requiredPermissions": {
            "description": "Required permissions for this plugin",
            "example": ["payments.read", "payments.write"],
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": [
          "name",
          "version",
          "description",
          "category",
          "bundleUrl",
          "extensionPoints",
          "adminComponents",
          "storefrontComponents",
          "metadata"
        ]
      },
      "PluginCategory": {
        "type": "string",
        "enum": [
          "payment",
          "notification",
          "analytics",
          "integration",
          "ui",
          "workflow"
        ],
        "description": "Plugin category"
      },
      "PluginStatus": {
        "type": "string",
        "enum": ["active", "inactive", "deprecated", "removed"],
        "description": "Plugin status"
      },
      "PluginResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "name": {
            "type": "string",
            "description": "Plugin name",
            "example": "Stripe Payment Gateway"
          },
          "version": {
            "type": "string",
            "description": "Plugin version",
            "example": "1.0.0"
          },
          "description": {
            "type": "string",
            "description": "Plugin description",
            "example": "Integrates Stripe payment processing"
          },
          "category": {
            "description": "Plugin category",
            "allOf": [{ "$ref": "#/components/schemas/PluginCategory" }]
          },
          "status": {
            "description": "Plugin status",
            "allOf": [{ "$ref": "#/components/schemas/PluginStatus" }]
          },
          "bundleUrl": {
            "type": "string",
            "description": "URL to the plugin bundle"
          },
          "extensionPoints": {
            "description": "List of extension points this plugin implements",
            "type": "array",
            "items": { "type": "string" }
          },
          "requiredPermissions": {
            "description": "Required permissions for this plugin",
            "type": "array",
            "items": { "type": "string" }
          },
          "metadata": {
            "type": "object",
            "description": "Metadata about the plugin",
            "example": {
              "priority": 100,
              "displayName": "Stripe Payments",
              "author": "Stripe Inc."
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "description",
          "category",
          "status",
          "extensionPoints",
          "requiredPermissions",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}